Use epoll
Ruby's thread scheduling gets messed up when enabling epoll for EM, so
the crash reaper needed to be refactored to make its specs pass.

The observed scheduling changes include that it may take an obscene
amount of time before a thread is scheduled (500ms or more).
To compensate for this, use EM.defer and wait for the work queue and
result queue to be empty before testing a spec's postcondition.

Change-Id: I6be18129b45b95d4d35feb5d5621301a0b6d5612