Use epoll
Ruby's thread scheduling gets messed up when enabling epoll for EM, so
the crash reaper needed to be refactored to make its specs pass.
Additionally, some time constrains needed to be relaxed.

The observed scheduling changes include that it may take an obscene
amount of time before a thread is scheduled (500ms or more).
To compensate for this, use EM.defer and wait for the work queue and
result queue to be empty before testing a spec's postcondition.

EventMachine hardcodes the epoll timeout to 50ms, which affects the
timer precision. Newer versions of EventMachine fix this, but this
repository is currently locked to a custom version of 0.12.10.

Change-Id: I6be18129b45b95d4d35feb5d5621301a0b6d5612