Use epoll
Ruby's thread scheduling gets messed up when enabling epoll for EM, so
a couple of things needed to be refactored.

Instead of using a custom thread and queue, use `EM.defer`.

Time constrains needed to be relaxed.

EventMachine hardcodes the epoll timeout to 50ms, which affects the
timer precision (newer versions of EventMachine fix this, but this
repository is currently locked to a custom version of 0.12.10).
This makes it hardly possible for threads other than the reactor threads
to run. To increase the probability that other threads also make
progress, `Thread.pass` is called from a timer that approximately runs
in every tick.

Change-Id: I6be18129b45b95d4d35feb5d5621301a0b6d5612