Revert "Don't map port when the application doesn't have URIs"
This reverts commit 89c70a52827a12b51f54c0fec52c65170ee5c7f4.

During the creation of an app instance, we would always like to have a
port assigned to the instance, even during the absence of URLs for the
app. Advantage is that we can 'vmc push' the app without a URL and
then 'vmc map' a URL to the app. Though we are leaking ports to
non web apps or web apps that will never ever have URLs, the former
case is negligible and the latter case is unlikely to happen.

Test plan:
 - All unit tests passed.
 - No BVT regression.
 - Manually tested the change:
   - Created an app.
   - Pushed it without an URL using 'vmc push'.
   - Mapped a URL to it using 'vmc map'.
   - The mapped URL was accessible.

Change-Id: I530ba5a98976000e2dea770fedfa69579c4c90a7