Update health check logic
Previously, health checks would try to connect to apps'
ports if they had been assigned an 'instance_host_port',
which would only happen if they had URIs mapped to them.
As of this commit: 15d52efc31206, apps are always assigned
an instance_host_port, regardless of how many URIs are mapped
to it.

Now, health checks are only done if there are URIs mapped
to the app.

Change-Id: I250b2f9dce8156f568cf41ae24e3ef722f8d29d2