Clone a single revision
Previously, when a buildpack was cloned, the entire buildpack
repository was fetched.  This resulted in a slower clone than was
otherwise necessary.  This change executes the clone with a --depth 1
which only fetches a single revision and no history.  This results in
noticeable speed gains.

Buildpack                       Full    Depth 1
===============================================
cloudfoundry-buildpack-java:    2.031s  1.451s
heroku-buildpack-nodejs:        6.840s  1.577s
heroku-buildpack-ruby:          5.075s  1.785s

[#50080879]