update running version if included in dea.update
This allows for a dea.update to be missed, and the system will
eventually resolve it.

If a DEA misses this message, it'll be hearbeating the old version. The
system converges to the current version, which means the route update
will eventually take effect (by killing the old instances and bringing
up new ones, with the correct routes).

There is a corresponding change in the CC, to update an app's version
when its routes are modified, and include the new version in the
dea.update that it sends.

[#57361124]