Temporary workaround for flaky integration spec
This is not a good long-term fix, but some tests depended on files being
around for longer than 10 seconds.  It's the test that is flaky, and we
will need to fix that test (running_the_app_spec, 'actually stops the
app') when we are under less pressure to fix the build pipeline.