Expose available_disk in advertisement message.
Cloud Controller currently assumes that if a DEA has enough memory, it
has enough disk. This is not always true (especially if an app requires
a large disk_quota). This commit adds the available_disk to the message
so that Cloud Controller can use this information to check if a DEA has
space before attempting to start the app.