Consistently use bytes for memory statistics
For some reason, stat_collector assumed rss was reported in kilobytes
instead of bytes and multiplied it by 1024.  The rest of the code then
started to divide the 'used_memory_in_bytes' by 1024 to really get bytes
back.  Very odd.

This change simply removes the hacks that multiply and divide to get
expected data.