Emit staging messages even when staging fails
Prior to loggregator, staging output was provided to the user by tailing
the logs generated during staging.  This pattern enabled end users to
see messages generated while staging was in-flight and included messages
related to ffailures in staging, but the newer Loggregator paths only emit
staging messages when staging completes successfully.

This change will cause staging messages to be emitted even when staging
fails but it still requires the staging task to complete.  An
investigation is needed into whether or not (and how) to emit messages
while staging is occuring as that seems like the 'correct' approach.