Provide config option for DEA to disable droplet cache
At present, when starting application from a DEA component, the app's droplet
file will be downloaded and cached in DEA. This mechanism will lead to some
issues in some resource restriction environments, as for some cases, the DEA
has a finite amount of disk capacity and if DEA caches the droplet files, the
disk space will run out. Also, for some security reasons, we don't want
everyone's app bits kept in the DEA.

As a result, it should be possible to disable the app bits cache in such
resource constrained and security sensitive environments. This patch provides
a config option for DEA named "disable_droplet_cache" to resolve this issue.
By default, this config option is false, when it is set true, DEA will remove
the droplet file cached in the base_dir/droplet/ folder after the instance of
an app is started.